Refactor debug to logger and update class names
