+++ packages/adk/src/memory/persistent-memory-service.ts
@@ -1,6 +1,6 @@
 import fs from "node:fs";
 import path from "node:path";
-import { Logger } from "@adk/helpers/debug";
+import { Logger } from "@adk/helpers/logger";
 import type { Session } from "../sessions/session";
 import type {
 	BaseMemoryService,
