+++ packages/adk/src/models/llm-registry.ts
@@ -1,4 +1,4 @@
-import { Logger } from "@adk/helpers/debug";
+import { Logger } from "@adk/helpers/logger";
 import type { BaseLLM } from "./base-llm";
 
 /**
